chore: move RealmConfigArgsV1 and GovernanceInstructionV1 to tests
